NOT_UPSTREAM: KVM: x86: Print SPTEs at each level of TDP page table for a given GPA
To test recovery from a Machine Check due to hardware page walk on a
hardware-poisoned EPT table, it is necessary to inject an uncorrectable
memory error into the EPT page table.

To provide the physical address of an EPT page table for this kind of
error injection, add a KVM debugfs node that prints SPTEs at each level
of the TDP page table (EPT table for Intel). The non-leaf SPTEs contain
the physical addresses of the EPT page table pages.

The following debugfs node, where the 0x110ed5000 is an example of GPA

  sudo chmod go+rwx /sys/kernel/debug
  echo 0x110ed5000 > /sys/kernel/debug/kvm/<kvm ID>/print_sptes

Print SPTEs to dmesg log such as:

  [  526.389237] kvm: GPA 110ed5000, spte[5] = 0x80000030a174e907, spte[4] = 0x80000030a174d907, spte[3] = 0x80000030c8a60907, spte[2] = 0x8600003118a00bf7

  The non-leaf SPTEs contains the physical addresses of the related EPT page table pages:

  spte[5] points to a level-4 EPT page table at physical address 0x30a174e000.
  spte[4] points to a level-3 EPT page table at physical address 0x30a174d000.
  spte[3] points to a level-2 EPT page table at physical address 0x30c8a60000.

  Any of the three physical addresses above can be used as input for error
  injection to test hardware-poisoned EPT recovery.

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>