stop_machine: Add stop_cluster_cpuslocked() for per-cluster operations
There is already stop_core_cpuslocked() for hardware core-level testing
features, which requires near-simultaneous execution of the WRMSR
instruction on all threads of a core to initiate a test. However, some
hardware requires this simultaneous execution of WRMSR at the cluster level.

Add stop_cluster_cpuslocked(), a customized version of
stop_machine_cpuslocked() for operating solely on the threads of a
single cluster.

Suggested-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>