platform/x86/intel/ifs: Revert SBAF feature
Structural Based Functional Test at Field (SBAF) test was initially
expected to be supported in Granite Rapids (GNR) CPUs. However, due to
recent roadmap changes, the SBAF test support will not be included in
GNR and is unlikely to be included in future CPUs.

Additionally, Intel has never published the SBAF scan files, and the
SBAF feature and its interface [1] are not utilized. Therefore, it's
safe to remove them. Given this, remove the SBAF test from the In-Field
Scan (IFS) driver by reverting the following commits:

  commit 1e701372d7ac ("platform/x86/intel/ifs: Fix SBAF title underline length")
  commit 61b74964536e ("trace: platform/x86/intel/ifs: Add SBAF trace support")
  commit 3c4d06bd6e37 ("platform/x86/intel/ifs: Add SBAF test support")
  commit 0a3e4e94d137 ("platform/x86/intel/ifs: Add SBAF test image loading support")
  commit 744a6a1f2a83 ("tools arch x86: Sync the msr-index.h copy with the kernel sources")

[1] /sys/devices/virtual/misc/intel_ifs_2

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>